Remove codel duplication of SpectrumDataset
